GeoJson of Nepal constituencies for 2023 elections
A long process but worth the pain

When we started working on Nepal election, the first requirement was a map of the constituencies to display the election results. The ultimate goal was to get the shape file with all constituencies.

Like every other person our first response was to google but we ended up getting nothing. The Election Commission of Nepal also did not publish the constituency map but they did have maps of each districts with the constituencies marked on it.

After many hit and trials QGIS came to our rescue. We worked on QGIS to make the constituency map for Nepal(which I think, didn't exist till the time we published it).

And voilà we had our Eureka moment after many tiring days. 

You can check BBC Nepali service page to check their election coverage.